+++ common/extension/config_center.go
@@ -18,6 +18,10 @@
 package extension
 
 import (
+	"errors"
+)
+
+import (
 	"dubbo.apache.org/dubbo-go/v3/common"
 	"dubbo.apache.org/dubbo-go/v3/config_center"
 )
@@ -31,8 +35,13 @@ func SetConfigCenter(name string, v func(*common.URL) (config_center.DynamicConf
 
 // GetConfigCenter finds the DynamicConfiguration with @name
 func GetConfigCenter(name string, config *common.URL) (config_center.DynamicConfiguration, error) {
-	if configCenters[name] == nil {
-		panic("config center for " + name + " is not existing, make sure you have import the package.")
+	configCenterFactory := configCenters[name]
+	if configCenterFactory == nil {
+		return nil, errors.New("config center for " + name + " is not existing, make sure you have import the package.")
 	}
-	return configCenters[name](config)
+	configCenter, err := configCenterFactory(config)
+	if err != nil {
+		return nil, err
+	}
+	return configCenter, nil
 }