+++ config/config_utils_test.go
@@ -61,3 +61,23 @@ func TestRemoveMinus(t *testing.T) {
 	strList = removeMinus([]string{"c", "b", "a", "d", "c", "-c", "-a", "e", "f"})
 	assert.Equal(t, strList, "b,d,c,e,f")
 }
+
+type mockConfig struct {
+	protocol string
+	address  string
+}
+
+func (m *mockConfig) Prefix() string {
+	return "mock"
+}
+
+func TestClientNameID(t *testing.T) {
+	t.Run("normal", func(t *testing.T) {
+		m := &mockConfig{"nacos", "127.0.0.1:8848"}
+		assert.Equal(t, "mock-nacos-127.0.0.1:8848", clientNameID(m, m.protocol, m.address))
+	})
+	t.Run("protocolIsNil", func(t *testing.T) {
+		m := &mockConfig{}
+		assert.Equal(t, "mock--", clientNameID(m, m.protocol, m.address))
+	})
+}