+++ config/root_config.go
@@ -85,6 +85,9 @@ type RootConfig struct {
 	CacheFile string `yaml:"cache_file" json:"cache_file,omitempty" property:"cache_file"`
 
 	Custom *CustomConfig `yaml:"custom" json:"custom,omitempty" property:"custom"`
+
+	// Profiles config
+	Profiles *ProfilesConfig `yaml:"profiles" json:"profiles,omitempty" property:"profiles"`
 }
 
 func SetRootConfig(r RootConfig) {
@@ -406,10 +409,11 @@ func (rc *RootConfig) Process(event *config_center.ConfigChangeEvent) {
 		logger.Errorf("CenterConfig process unmarshalConf failed, got error %#v", err)
 		return
 	}
-
-	// update register
+	// dynamically update register
 	for registerId, updateRegister := range updateRootConfig.Registries {
 		register := rc.Registries[registerId]
-		register.UpdateProperties(updateRegister)
+		register.DynamicUpdateProperties(updateRegister)
 	}
+	// dynamically update consumer
+	rc.Consumer.DynamicUpdateProperties(updateRootConfig.Consumer)
 }